+++ ELEMNTViewer/RibbonMarkup1.h
@@ -1,0 +1,41 @@
+// *****************************************************************************
+// * This is an automatically generated header file for UI Element definition  *
+// * resource symbols and values. Please do not modify manually.               *
+// *****************************************************************************
+
+#pragma once
+
+#define cmdTabStart 2 
+#define cmdTabStart_LabelTitle_RESID 60001
+#define cmdComboYear 3 
+#define cmdComboYear_LabelTitle_RESID 60002
+#define cmdComboYear_TooltipTitle_RESID 60003
+#define cmdComboYear_TooltipDescription_RESID 60004
+#define cmdComboMonth 4 
+#define cmdComboMonth_LabelTitle_RESID 60005
+#define cmdComboMonth_TooltipTitle_RESID 60006
+#define cmdComboMonth_TooltipDescription_RESID 60007
+#define cmdGroupSelect 5 
+#define cmdGroupSelect_LabelTitle_RESID 60008
+#define cmdGroupSelectFunction 6 
+#define cmdGroupSelectFunction_LabelTitle_RESID 60009
+#define cmdGroupSelect2 11 
+#define cmdGroupSelect2_LabelTitle_RESID 60010
+#define cmdCheckMonthly 12 
+#define cmdCheckMonthly_LabelTitle_RESID 60011
+#define cmdCheckMonthly_TooltipTitle_RESID 60012
+#define cmdCheckMonthly_TooltipDescription_RESID 60013
+#define cmdCheckAntDevice 9 
+#define cmdCheckAntDevice_LabelTitle_RESID 60014
+#define cmdCheckAntDevice_TooltipTitle_RESID 60015
+#define cmdCheckAntDevice_TooltipDescription_RESID 60016
+#define cmdGroupAnt 10 
+#define cmdGroupAnt_LabelTitle_RESID 60017
+#define cmdComboAnt 13 
+#define cmdComboAnt_LabelTitle_RESID 60018
+#define cmdComboAnt_TooltipTitle_RESID 60019
+#define cmdComboAnt_TooltipDescription_RESID 60020
+#define cmdComboAntName 14 
+#define cmdComboAntName_TooltipTitle_RESID 60021
+#define cmdComboAntName_TooltipDescription_RESID 60022
+#define InternalCmd2_LabelTitle_RESID 60023