+++ ELEMNTViewer/app/AntDevice.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ELEMNTViewer
+{
+    struct AntDevice
+    {
+        public ushort Id { get; set; }
+        public string ProductName { get; set; }
+    }
+}